refactored snippetPreview, moved it into a plugin folder
