+++ composites/Plugin/SnippetPreview/components/Button.js
@@ -1,6 +1,6 @@
 import styled from "styled-components";
-import colors from "../../../style-guide/colors.json";
-import { rgba } from "../../../style-guide/helpers";
+import colors from "../../../../style-guide/colors.json";
+import { rgba } from "../../../../style-guide/helpers";
 
 /**
  * Returns a component with applied focus styles.
