+++ composites/Plugin/SnippetPreview/components/Icon.js
