+++ composites/Plugin/SnippetPreview/tests/IconTest.js
@@ -1,7 +1,7 @@
 import React from "react";
 import renderer from "react-test-renderer";
 
-import edit from "../../../style-guide/svg/edit.svg";
+import edit from "../../../../style-guide/svg/edit.svg";
 import { Icon } from "../components/Icon";
 
 test( "the Icon matches the snapshot", () => {
